backend/requirements.txt
# Core Flask
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0

# Google APIs
google-api-python-client>=2.110.0
google-generativeai>=0.3.2

# YouTube
youtube-transcript-api>=0.6.2
isodate>=0.6.1

# Database
supabase>=2.3.0

# Environment
python-dotenv>=1.0.0

# Optional: ML libraries (only needed for advanced analytics)
# Uncomment if you need them and have a C compiler installed
# numpy>=1.26.0
# scipy>=1.11.0
# scikit-learn>=1.4.0
pypdf>=4.0.0
